project_03_langgraph_research/requirements.txt
# LangGraph
langgraph>=0.2.0

# Google AI (direct, no LangChain wrapper to avoid transformers dependency)
google-generativeai

# LangChain for tools only
langchain>=0.3.0
langchain-core
langchain-community

# Search tool
langchain-tavily
tavily-python

# Environment
python-dotenv

# Utilities
pydantic>=2.0.0
typing-extensions
